Narrow the core extension boundary
`@optique/core` should not expose parser markers, replay state, and
annotation transport details as supported API because that makes
internal refactors semver-sensitive. Move those details behind internal
modules, keep the supported low-level helpers under
`@optique/core/extension`, and update first-party integrations and docs
to follow the new boundary.

Fixes https://github.com/dahlia/optique/issues/792
