+++ packages/core/deno.json
@@ -13,7 +13,6 @@
     "./extension": "./src/extension.ts",
     "./facade": "./src/facade.ts",
     "./message": "./src/message.ts",
-    "./mode-dispatch": "./src/mode-dispatch.ts",
     "./modifiers": "./src/modifiers.ts",
     "./nonempty": "./src/nonempty.ts",
     "./parser": "./src/parser.ts",
