+++ packages/core/src/dependency-runtime.test.ts
@@ -32,9 +32,9 @@ import {
   DependencyRegistry,
   isDependencySourceState,
   parseWithDependency,
-} from "./dependency.ts";
+} from "./internal/dependency.ts";
 import { message } from "./message.ts";
-import { unmatchedNonCliDependencySourceStateMarker } from "./parser.ts";
+import { unmatchedNonCliDependencySourceStateMarker } from "./internal/parser.ts";
 import type { DependencyRegistryLike } from "./registry-types.ts";
 import type { ValueParserResult } from "./valueparser.ts";
