+++ packages/core/tsdown.config.ts
@@ -12,7 +12,6 @@ export default defineConfig({
     "src/extension.ts",
     "src/facade.ts",
     "src/message.ts",
-    "src/mode-dispatch.ts",
     "src/modifiers.ts",
     "src/nonempty.ts",
     "src/parser.ts",
